adding client-side redirect if no user
it seems like the preloading of the profile page and other routes is
able to bypass the hooks.server authguard...

need to look into this. there is a code version to preload data and an
html attribute (current impl). may need to switch to code version to
force check through hooks.server... but that may void the perf gain of
the preloading lol...